Add a submission year.
This commits introduces a pragmatic solution to the fact that we now
have submissions from multiple years, without a real submission time
or year being included in the .xml files.

It seems rather "unsafe" to ask students to provide that submission
year anyway, so we force it at import time.

Now, a submitted_at attribute already existed, which is not really
used so far. We will keep both 'at' and 'year' for now and clean it
later.